requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.4.0
alembic==1.16.1
altair==5.5.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.9.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==14.4.0
babel==2.17.0
bangla==0.0.5
bidict==0.23.1
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
CacheControl==0.14.3
cachetools==6.1.0
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
colorama==0.4.6
comtypes==1.4.11
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cryptography==45.0.3
cvzone==1.6.1
cycler==0.12.1
cymem==2.0.11
Cython==3.1.2
dateparser==1.1.8
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
einops==0.8.1
en_core_web_sm @ file:///C:/Users/atsha/invoice_sap_project/en_core_web_sm-3.8.0-py3-none-any.whl
encodec==0.1.1
exceptiongroup==1.3.0
filelock==3.18.0
firebase_admin==7.1.0
flasgger==0.9.7.1
Flask==2.3.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.2
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.58.0
fpdf==1.7.2
frozenlist==1.7.0
fsspec==2025.5.1
g2pkk==0.1.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.2
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.75.1
grpcio-status==1.75.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
h11==0.16.0
h2==4.3.0
h5py==3.13.0
hangul-romanize==0.1.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
ifaddr==0.2.0
imbalanced-learn==0.13.0
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jax==0.6.0
jaxlib==0.6.0
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonlines==1.2.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.10.0
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.0
llvmlite==0.43.0
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mediapipe==0.10.21
mistune==3.1.3
ml_dtypes==0.5.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
murmurhash==1.0.13
namex==0.0.9
narwhals==2.1.2
networkx==2.8.8
nltk==3.8.1
num2words==0.5.14
numba==0.60.0
numpy==1.24.0
openai==0.28.1
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.3.1
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.9.0
pillow==10.3.0
platformdirs==4.3.8
playsound==1.2.2
plotly==6.3.1
pooch==1.8.2
preshed==3.0.9
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
pyenchant==3.2.2
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.12.0
pymongo==4.3.3
pyngrok==7.3.0
pynndescent==0.5.13
pyOpenSSL==25.1.0
pyparsing==3.0.9
PyPDF2==3.0.1
pypdfium2==4.30.1
pypinyin==0.54.0
pypiwin32==223
pysbd==0.3.4
pytesseract==0.3.13
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.12.2
python-socketio==5.13.0
python-vlc==3.0.21203
pyttsx3==2.98
pytz==2025.2
pywin32==310
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==5.0.0
sentencepiece==0.2.1
shellingham==1.5.4
simple-websocket==1.1.0
simpleaudio==1.0.4
six==1.17.0
sklearn-compat==0.1.3
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.7
spacy-alignments==0.9.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.9
SpeechRecognition==3.10.0
SQLAlchemy==2.0.41
srsly==2.5.1
streamlit==1.48.1
streamlit-webrtc==0.63.4
SudachiDict-core==20250515
SudachiPy==0.6.10
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tf_keras==2.19.0
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.8.0
torchaudio==2.7.1
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
trainer==0.0.36
transformers==4.55.2
TTS==0.22.0
twilio==9.6.5
typeguard==4.4.4
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.4.0
urllib3==2.4.0
Wand==0.6.13
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
Werkzeug==2.3.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
yt-dlp==2025.7.21
